优化文件架构和代码风格
