--- swarmclone/Frontend/__init__.py
@@ -1,1 +1,0 @@
-from .Frontend import frontend
