--- swarmclone/Frontend2/__init__.py
@@ -1,1 +1,0 @@
-from .Frontend2 import frontend2
