+++ swarmclone/__main__.py
@@ -2,11 +2,11 @@ from .controller import Controller
 from .modules import *
 from .constants import *
 from .tts_cosyvoice import TTSCosyvoice
-from .Frontend import frontend
+from .frontend_socket import FrontendSocket
 
 if __name__ == "__main__":
     controller = Controller()
-    controller.register(frontend())
+    controller.register(FrontendSocket())
     controller.register(TTSCosyvoice())
     controller.register(LLMDummy())
     controller.register(ASRDummy())
