+++ swarmclone/frontend_socket.py
@@ -2,14 +2,14 @@ import asyncio
 import json
 import base64
  
-from swarmclone.config import config
-from ..modules import ModuleRoles, ModuleBase
-from ..messages import *
+from .config import config
+from .modules import ModuleRoles, ModuleBase
+from .messages import *
 
-class frontend(ModuleBase):
+class FrontendSocket(ModuleBase):
     def __init__(self):
-        super().__init__(ModuleRoles.FRONTEND, "Frontend")
-        self.clientdict:dict[int,asyncio.StreamWriter] = {}
+        super().__init__(ModuleRoles.FRONTEND, "FrontendSocket")
+        self.clientdict: dict[int, asyncio.StreamWriter] = {}
         self.server = None
 
     async def run(self):
@@ -20,8 +20,8 @@ class frontend(ModuleBase):
             await self.server.serve_forever()
     
     async def preprocess_tasks(self):
-        while(True):
-            if(self.clientdict):
+        while True:
+            if self.clientdict:
                 task = await self.task_queue.get()
                 to_remove = []
                 message = self.load(task)
@@ -42,19 +42,21 @@ class frontend(ModuleBase):
     def handle_client(self, reader:asyncio.StreamReader, writer:asyncio.StreamWriter) -> None:
         addr = writer.get_extra_info('peername')
         print(f"客户端已连接：{addr}")
-        self.clientdict[addr[1]]=writer
+        self.clientdict[addr[1]] = writer
     
-    def load(self, task:Message) -> str:
+    def load(self, task: Message) -> str:
         dict = {
-                "message_type": task.message_type.value,
-                "source": task.source.role.value,
-                **task.kwargs
-            }
-        if(isinstance(task, TTSAudio)):
+            "message_type": task.message_type.value,
+            "source": task.source.role.value,
+            **task.kwargs
+        }
+        if isinstance(task, TTSAudio):
             dict["data"] = base64.b64encode(dict["data"]).decode('utf-8')#UnicodeDecodeError: 'utf-8' codec can't decode byte 0x81 in position 1: invalid start byte
-        massage = (json.dumps(dict).replace(config.panel.server.requests_separator, "") + # 防止在不应出现的地方出现分隔符
-        config.panel.server.requests_separator)
-        return(massage)
+        massage = (
+            json.dumps(dict).replace(config.panel.server.requests_separator, "") + # 防止在不应出现的地方出现分隔符
+            config.panel.server.requests_separator
+        )
+        return massage
     
     async def process_task(self, task: Message | None) -> Message | None:
     # 不应被调用
